+++ src/test/java/com/solvd/blog/transaction/UsersIT.java
@@ -1,6 +1,5 @@
-package com.solvd.blog.neo4j;
+package com.solvd.blog.transaction;
 
-import com.solvd.blog.exception.ResourceAlreadyExistsException;
 import com.solvd.blog.fake.FkUser;
 import com.solvd.blog.model.User;
 import com.solvd.blog.model.Users;
@@ -18,7 +17,7 @@ import org.springframework.test.context.DynamicPropertySource;
 public class UsersIT extends Neo4jIntegration {
 
     @Autowired
-    @Qualifier("vdUsers")
+    @Qualifier("txUsers")
     private Users users;
 
     @DynamicPropertySource
@@ -60,14 +59,6 @@ public class UsersIT extends Neo4jIntegration {
     }
 
     @Test
-    public void verifiesAddThrowsResourceAlreadyExistsException() {
-        Assertions.assertThrows(
-                ResourceAlreadyExistsException.class,
-                () -> this.users.add(new FkUser("Name", "Email"))
-        );
-    }
-
-    @Test
     public void verifiesUpdate() {
         User expected = new FkUser("Updated name", "Email");
         User result = this.users.update(expected);
@@ -76,7 +67,7 @@ public class UsersIT extends Neo4jIntegration {
     }
 
     @Test
-    public void verifiesIsExists() {
+    public void verifiesExists() {
         Assertions.assertTrue(
                 this.users.exists("Email")
         );
